% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{alluvial_flow.plot}
\alias{alluvial_flow.plot}
\title{Alluvial flow diagram}
\usage{
alluvial_flow.plot(dtaMatrix)
}
\arguments{
\item{datMatrix}{jaccard indecies matrix}
}
\value{
alluvial diagram plot
}
\description{
Alluvial flow diagram
}
