% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{apply_kmeans}
\alias{apply_kmeans}
\title{apply kmeans clustering}
\usage{
apply_kmeans(dataMatrix, k)
}
\arguments{
\item{dataMatrix}{an counts matrix, after pre-processing and dimensionality reduction}

\item{k}{the number of clusters}
}
\value{
clustering labels
}
\description{
apply kmeans clustering
}
