% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{apply_SC3}
\alias{apply_SC3}
\title{apply SC3 clustering methodology}
\usage{
apply_SC3(datMatrix, addProc, k = NULL)
}
\arguments{
\item{datMatrix}{a gene counts matrix after pre-processing}

\item{addProc}{a logical parameter for additional processing parameter in SC3}
}
\value{
clustering labels
}
\description{
apply SC3 clustering methodology
}
\examples{
eg1:
SC3_estimate_results <- apply_SC3(normalized,
addProc = TRUE)

eg2:
SC3_estimate_resutls <- apply_SC3(normalized,
addProc = FALSE)

eg3:
SC3_estimate_resutls <- apply_SC3(normalized,
k = 8)

}
