% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_performance.R
\name{balancedAccuracy}
\alias{balancedAccuracy}
\title{Measure balanced accuracy}
\usage{
balancedAccuracy(recall, specificity)
}
\arguments{
\item{recall}{is the recall value}

\item{specificty}{is the specificity value}
}
\value{
balanced accuracy value
}
\description{
Measure balanced accuracy
}
