% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_performance.R
\name{bench_process}
\alias{bench_process}
\title{Benchmark with ARI and NMI metrics}
\usage{
bench_process(labels, clusters, prf.metrics)
}
\arguments{
\item{labels}{data ground truth labels}

\item{clusters}{resulted clusters}
}
\value{
benchmark barplot fig
}
\description{
Benchmark with ARI and NMI metrics
}
