% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{benchmark_barplot}
\alias{benchmark_barplot}
\title{Barplot for benchmark}
\usage{
benchmark_barplot(dta, pal = barpalette2)
}
\arguments{
\item{dta}{dataframe of all package metrics alongside ARI and NMI values}

\item{pal}{colour palette "must be more than 7 colours" (deafult = barpalette2)}
}
\value{
the barplot fig
}
\description{
Barplot for benchmark
}
