% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{bxplot_metric}
\alias{bxplot_metric}
\title{Visualize clustering performance by specific metric}
\usage{
bxplot_metric(lngform, metric = "F1-Score", pal = gg_bxpalette)
}
\arguments{
\item{lngform}{melted dataframe of performance metrics}

\item{metric}{chosen performance metric, default = F1-Score}
}
\value{
box plot visualization of the chosen metric
}
\description{
Visualize clustering performance by specific metric
}
