% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionalities.R
\name{calc_confusion_matrix}
\alias{calc_confusion_matrix}
\title{Confusion Matrix Metrics}
\usage{
calc_confusion_matrix(sample, labels, clusters)
}
\arguments{
\item{sample}{is the sample name to search by}

\item{labels}{is the list of samples true labels}

\item{clusters}{is the list of samples assigned clusters}
}
\value{
whether the entered sample is a TP, FP or FN
}
\description{
calculate the confusion matrix for each entered samples by comparing
lables against clusters
}
