% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{confusionMatrix_plot}
\alias{confusionMatrix_plot}
\title{Build confusion matrix plot}
\usage{
confusionMatrix_plot(datMatrix)
}
\arguments{
\item{datMatrix}{matrix of size= labels x clusters}
}
\value{
confusion matrix figure
}
\description{
Build confusion matrix plot
}
