% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_performance.R
\name{F1Score}
\alias{F1Score}
\title{Measure F1-Score}
\usage{
F1Score(prec, rec)
}
\arguments{
\item{rec}{is the recall value}

\item{pres}{is the precision value}
}
\value{
F1-Score value
}
\description{
Measure F1-Score
}
