% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{lb_clust_plot}
\alias{lb_clust_plot}
\title{Scatter plot for labels and clusters}
\usage{
lb_clust_plot(mtx, labels, fgTitle = "", lgndPos = "none")
}
\arguments{
\item{mtx}{a dataframe of low dimensional components}

\item{labels}{cell types or clustering labels}

\item{lgndPos}{legend position (none by default)}
}
\value{
scatter plot colored by labels
}
\description{
Scatter plot for labels and clusters
}
