% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{longFormat}
\alias{longFormat}
\title{Melt dataframe}
\usage{
longFormat(dfList)
}
\arguments{
\item{dfList}{list of clustering performance dataframes}
}
\value{
re-formated shape suitable for multiple plotting techniques, eg: heatmap, boxplot, etc..
}
\description{
Melt dataframe
}
