% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_performance.R
\name{measurePerformance}
\alias{measurePerformance}
\title{Measure performance metrics}
\usage{
measurePerformance(pn)
}
\arguments{
\item{pn}{is the confusion matrix}

\item{k}{is the number of clusters}
}
\value{
a dataframe of all performance metrics calculated
}
\description{
This function measure the following performace metrics for each cluster and then
take the average of each metric along all clusters:
   a. Precision
   b. Recall
   c. Specificity
   d. Balanced Accuracy
   e. F1-Score
}
