% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{metrics_heatmap}
\alias{metrics_heatmap}
\title{Performance heatmap plot}
\usage{
metrics_heatmap(metric.df)
}
\arguments{
\item{melted.df}{dataframe formulated to build heatmap}
}
\value{
heatmap plot
}
\description{
Performance heatmap plot
}
