% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionalities.R
\name{multiclass_handling}
\alias{multiclass_handling}
\alias{handle_multiclass}
\title{Multi-class confusion matrix}
\usage{
handle_multiclass(samples, labels, clusters, maxmatch.mp)
}
\arguments{
\item{samples}{is a list of the samples names}

\item{labels}{is a list of samples true labels}

\item{clusters}{is a list of the samples assigned clusters}

\item{k}{is the number of clusters}

\item{mapped}{is a dataframe of the mapping results}
}
\value{
a dataframe of each class (TP, TN, FP, FN)
}
\description{
obtain the confusion matrix metrics for multi-class data using the
mapping results
}
