% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{pca}
\alias{pca}
\title{Dimensionality reduction using PCA}
\usage{
pca(datMatrix, norm = FALSE)
}
\arguments{
\item{datMatrix}{counts matrix}

\item{norm}{logical parameter to check whether normalized counts or not}
}
\value{
low dimensional components
}
\description{
Dimensionality reduction using PCA
}
