% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{performance.heatmap}
\alias{performance.heatmap}
\title{Performance heatmap plot}
\usage{
performance.heatmap(melted.df, datasetName)
}
\arguments{
\item{melted.df}{dataframe formulated to build heatmap}

\item{datasetName}{dataset name to add to title}
}
\value{
heatmap plot
}
\description{
Performance heatmap plot
}
