% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_performance.R
\name{precision}
\alias{precision}
\title{Measure precision}
\usage{
precision(TP, FP)
}
\arguments{
\item{TP}{true positive}

\item{FP}{false positive}
}
\value{
precision value
}
\description{
Measure precision
}
