% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{preprocess}
\alias{preprocess}
\title{Pre-processing}
\usage{
preprocess(datMatrix)
}
\arguments{
\item{data}{gene counts matrix}
}
\value{
a pre-processed counts matrix in sce object
}
\description{
Pre-processing
}
\examples{
\dontrun{
processed <- preprocess(sce)

# extract processed counts
processedCounts <- counts(sce)
}
}
