% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sankey.diagram}
\alias{sankey.diagram}
\alias{sankey.diag}
\title{Sankey diagram}
\usage{
sankey.diag(confusionMatrix, path)
}
\arguments{
\item{path}{file path to save sankey diagram in a .html format}

\item{confusioMatrix}{confusion matrix}
}
\description{
Sankey diagram
}
