% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_performance.R
\name{specificity}
\alias{specificity}
\title{Measure specificity}
\usage{
specificity(TN, FP)
}
\arguments{
\item{TN}{true negative}

\item{FP}{false positive}
}
\value{
specificity value
}
\description{
Measure specificity
}
